Task5_ProfilesRestApi/profiles-rest-api/requirements.txt
# Python = 3.6.1
Django==1.11
djangorestframework==3.6.2
pytz==2018.7
uWSGI==2.0.15
